ConnectionTcpObfuscated

Telegram obfuscated2 connection that encrypts every message with a randomly generated key using AES-CTR.

Super classes

telegramR::Connection -> telegramR::ObfuscatedConnection -> ConnectionTcpObfuscated

Public fields

obfuscated_io

ObfuscatedIO class for handling obfuscated I/O operations.

packet_codec

Packet codec set to AbridgedPacketCodec.

Methods

Public methods

ConnectionTcpObfuscated$clone()

Inherited methods

telegramR::Connection$connect()

telegramR::Connection$disconnect()

telegramR::Connection$initialize()

telegramR::Connection$is_connected()

telegramR::Connection$recv()

telegramR::Connection$send()

telegramR::Connection$to_string()

Method clone()

The objects of this class are cloneable with this method.

Usage

ConnectionTcpObfuscated$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.