DraftsIter

Telegram API type DraftsIter

Details

DraftsIter class

Super class

telegramR::RequestIter -> DraftsIter

Methods

Public methods

DraftsIter$new()

DraftsIter$load_next_chunk()

DraftsIter$get_init_future()

DraftsIter$clone()

Inherited methods

telegramR::RequestIter$.next()

telegramR::RequestIter$async_init()

telegramR::RequestIter$collect()

telegramR::RequestIter$reset()

Method new()

Constructor for the DraftsIter class.

Usage

DraftsIter$new(client, limit, entities = NULL)

Arguments

client

The Telegram client.

limit

The maximum number of drafts to retrieve.

entities

The list of entities for which to fetch the drafts.

Returns

A new DraftsIter object.

Method load_next_chunk()

Loads the next chunk of drafts.

Usage

DraftsIter$load_next_chunk()

Returns

A future that resolves to a list of drafts.

Method get_init_future()

Gets the initialization future.

Usage

DraftsIter$get_init_future()

Returns

A future that resolves to the initialization result.

Method clone()

The objects of this class are cloneable with this method.

Usage

DraftsIter$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.