EditExportedChatInviteRequest

Represents a request to edit an exported chat invite. This class inherits from TLRequest.

Super classes

telegramR::TLObject -> telegramR::TLRequest -> EditExportedChatInviteRequest

Public fields

CONSTRUCTOR_ID

The constructor ID.

SUBCLASS_OF_ID

The subclass ID.

Methods

Public methods

EditExportedChatInviteRequest$new()

EditExportedChatInviteRequest$resolve()

EditExportedChatInviteRequest$toDict()

EditExportedChatInviteRequest$bytes()

EditExportedChatInviteRequest$clone()

Inherited methods

telegramR::TLObject$.bytes()

telegramR::TLObject$.eq()

telegramR::TLObject$.ne()

telegramR::TLObject$.str()

telegramR::TLObject$from_reader()

telegramR::TLObject$pretty_format()

telegramR::TLObject$serialize_bytes()

telegramR::TLObject$serialize_datetime()

telegramR::TLObject$stringify()

telegramR::TLObject$to_dict()

telegramR::TLObject$to_json()

telegramR::TLRequest$read_result()

Method new()

Initialize the EditExportedChatInviteRequest object.

Usage

EditExportedChatInviteRequest$new(
  peer,
  link,
  revoked = NULL,
  expireDate = NULL,
  usageLimit = NULL,
  requestNeeded = NULL,
  title = NULL
)

Arguments

peer

The input peer.

link

The invite link.

revoked

Whether the invite is revoked (optional).

expireDate

The expiration date (optional).

usageLimit

The usage limit (optional).

requestNeeded

Whether request is needed (optional).

title

The title (optional).

Method resolve()

Resolve the request using client and utils.

Usage

EditExportedChatInviteRequest$resolve(client, utils)

Arguments

client

The client object.

utils

The utils object.

Method toDict()

Convert the object to a dictionary.

Usage

EditExportedChatInviteRequest$toDict()

Returns

A list representing the object.

Method bytes()

Serialize the object to bytes.

Usage

EditExportedChatInviteRequest$bytes()

Returns

A raw vector of bytes.

Method clone()

The objects of this class are cloneable with this method.

Usage

EditExportedChatInviteRequest$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.