GetStoriesArchiveRequest R6 class

GetStoriesArchiveRequest R6 class

GetStoriesArchiveRequest R6 class

Details

Request to get archived stories for a peer with pagination (offset_id, limit). Returns stories.Stories.

Super classes

telegramR::TLObject -> telegramR::TLRequest -> GetStoriesArchiveRequest

Public fields

CONSTRUCTOR_ID

Constructor identifier for this TL object.

SUBCLASS_OF_ID

Subclass identifier for this TL object.

peer

Field.

offset_id

Field.

limit

Field.

Methods

Public methods

GetStoriesArchiveRequest$new()

GetStoriesArchiveRequest$resolve()

GetStoriesArchiveRequest$to_list()

GetStoriesArchiveRequest$to_bytes()

GetStoriesArchiveRequest$clone()

Inherited methods

telegramR::TLObject$.bytes()

telegramR::TLObject$.eq()

telegramR::TLObject$.ne()

telegramR::TLObject$.str()

telegramR::TLObject$from_reader()

telegramR::TLObject$pretty_format()

telegramR::TLObject$serialize_bytes()

telegramR::TLObject$serialize_datetime()

telegramR::TLObject$stringify()

telegramR::TLObject$to_dict()

telegramR::TLObject$to_json()

telegramR::TLRequest$read_result()

Method new()

Initialize GetStoriesArchiveRequest

Usage

GetStoriesArchiveRequest$new(peer, offset_id, limit)

Arguments

peer

TypeInputPeer

offset_id

integer offset id

limit

integer limit

Returns

invisible self

Method resolve()

Resolve peer references

Convert high-level peer reference to input peer using client/utils.

Usage

GetStoriesArchiveRequest$resolve(client, utils)

Arguments

client

client with get_input_entity

utils

utils with get_input_peer

Returns

invisible self

Method to_list()

Convert to list

Usage

GetStoriesArchiveRequest$to_list()

Returns

list

Method to_bytes()

Serialize to raw TL bytes

Usage

GetStoriesArchiveRequest$to_bytes()

Returns

raw

Method clone()

The objects of this class are cloneable with this method.

Usage

GetStoriesArchiveRequest$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.