ToggleNoPaidMessagesExceptionRequest

R6 class representing a ToggleNoPaidMessagesExceptionRequest.

Details

This class handles toggling the no paid messages exception for a user.

Super classes

telegramR::TLObject -> telegramR::TLRequest -> ToggleNoPaidMessagesExceptionRequest

Methods

Public methods

ToggleNoPaidMessagesExceptionRequest$new()

ToggleNoPaidMessagesExceptionRequest$resolve()

ToggleNoPaidMessagesExceptionRequest$toDict()

ToggleNoPaidMessagesExceptionRequest$bytes()

ToggleNoPaidMessagesExceptionRequest$fromReader()

ToggleNoPaidMessagesExceptionRequest$clone()

Inherited methods

telegramR::TLObject$.bytes()

telegramR::TLObject$.eq()

telegramR::TLObject$.ne()

telegramR::TLObject$.str()

telegramR::TLObject$from_reader()

telegramR::TLObject$pretty_format()

telegramR::TLObject$serialize_bytes()

telegramR::TLObject$serialize_datetime()

telegramR::TLObject$stringify()

telegramR::TLObject$to_dict()

telegramR::TLObject$to_json()

telegramR::TLRequest$read_result()

Method new()

Initialize the ToggleNoPaidMessagesExceptionRequest.

Usage

ToggleNoPaidMessagesExceptionRequest$new(
  userId,
  refundCharged = NULL,
  requirePayment = NULL,
  parentPeer = NULL
)

Arguments

userId

The input user ID.

refundCharged

Optional boolean for refund charged.

requirePayment

Optional boolean for require payment.

parentPeer

Optional input parent peer.

Method resolve()

Resolve the user ID and parent peer using client and utils.

Usage

ToggleNoPaidMessagesExceptionRequest$resolve(client, utils)

Arguments

client

The client object.

utils

The utils object.

Method toDict()

Convert to dictionary.

Usage

ToggleNoPaidMessagesExceptionRequest$toDict()

Returns

A list representing the object.

Method bytes()

Serialize to bytes.

Usage

ToggleNoPaidMessagesExceptionRequest$bytes()

Returns

Raw bytes.

Method fromReader()

Create from reader.

Usage

ToggleNoPaidMessagesExceptionRequest$fromReader(reader)

Arguments

reader

The reader object.

Returns

An instance of ToggleNoPaidMessagesExceptionRequest.

Method clone()

The objects of this class are cloneable with this method.

Usage

ToggleNoPaidMessagesExceptionRequest$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.